requirements.txt
dj-database-url==0.5.0
Django==3.2.7
django-extensions==3.1.3
djangorestframework==3.12.4
psycopg2==2.9.1
python-decouple==3.4
pre-commit==2.15.0
